+++ examples/demo_regression_perpendicular.py
@@ -4,7 +4,7 @@ from sklearn.metrics import mean_squared_error
 from bayesian_decision_tree.regression import PerpendicularRegressionNode
 from examples import helper
 
-# demo script for regression
+# demo script for regression using classic, axis-normal splits
 if __name__ == '__main__':
     # proxies (in case you're running this behind a firewall)
     args = helper.parse_args()
@@ -15,14 +15,14 @@ if __name__ == '__main__':
 
     # data set: uncomment one of the following sections
 
-    # synthetic sine wave
-    X_train = np.linspace(0, 10, 100).reshape(-1, 1)
-    y_train = 1 * np.sin(np.linspace(0, 10, 100)).reshape(-1, 1)
-    train = np.hstack((X_train, y_train))
-    test = train
+    # # synthetic sine wave
+    # X_train = np.linspace(0, 10, 100).reshape(-1, 1)
+    # y_train = 1 * np.sin(np.linspace(0, 10, 100)).reshape(-1, 1)
+    # train = np.hstack((X_train, y_train))
+    # test = train
 
-    # or, alternatively, load a UCI dataset
-    # train, test = helper.load_ripley(proxies)
+    # or, alternatively, load a UCI dataset (where we *regress* on the class labels, i.e., class 1 = 0.0 and class 2 = 1.0)
+    train, test = helper.load_ripley(proxies)
 
     n_dim = len(np.unique(train[:, -1]))
 
@@ -39,9 +39,9 @@ if __name__ == '__main__':
     # prior for regression: Normal-Gamma prior, see https://en.wikipedia.org/wiki/Conjugate_prior#Continuous_distributions
     mu = y_train.mean()
     sd_prior = y_train.std() / 10
-    prior_obs = 1
-    kappa = prior_obs
-    alpha = prior_obs/2
+    prior_pseudo_observations = 1
+    kappa = prior_pseudo_observations
+    alpha = prior_pseudo_observations / 2
     var_prior = sd_prior**2
     tau_prior = 1/var_prior
     beta = alpha/tau_prior
@@ -52,18 +52,18 @@ if __name__ == '__main__':
     delta = 0
 
     # model
-    root = PerpendicularRegressionNode(partition_prior, prior)
+    model = PerpendicularRegressionNode(partition_prior, prior)
 
     # train
-    root.fit(X_train, y_train, delta)
-    print(root)
+    model.fit(X_train, y_train, delta)
+    print(model)
     print()
-    print('Tree depth and number of leaves:', root.depth_and_leaves())
-    print('Feature importance:', root.feature_importance())
+    print('Tree depth and number of leaves:', model.depth_and_leaves())
+    print('Feature importance:', model.feature_importance())
 
     # compute RMSE
-    rmse_train = np.sqrt(mean_squared_error(root.predict(X_train), y_train))
-    rmse_test = np.sqrt(mean_squared_error(root.predict(X_test), y_test))
+    rmse_train = np.sqrt(mean_squared_error(model.predict(X_train), y_train))
+    rmse_test = np.sqrt(mean_squared_error(model.predict(X_test), y_test))
     info_train = 'RMSE train: {:.4f}'.format(rmse_train)
     info_test = 'RMSE test:  {:.4f}'.format(rmse_test)
     print(info_train)
@@ -72,6 +72,6 @@ if __name__ == '__main__':
     # plot if 1D or 2D
     dimensions = X_train.shape[1]
     if dimensions == 1:
-        helper.plot_1d_perpendicular(root, X_train, y_train, info_train, X_test, y_test, info_test)
+        helper.plot_1d_perpendicular(model, X_train, y_train, info_train, X_test, y_test, info_test)
     elif dimensions == 2:
-        helper.plot_2d_perpendicular(root, X_train, y_train, info_train, X_test, y_test, info_test)
+        helper.plot_2d_perpendicular(model, X_train, y_train, info_train, X_test, y_test, info_test)